# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: backend
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
...
  /projects/{project_id}/podcasts:
    get:
      tags: []
      operationId: list_podcasts
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      - name: cursor_created_at
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: cursor_podcast_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          format: int32
          maximum: 32
          minimum: 1
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, name, description, user_id, project_id, created_at, status]
                      properties:
                        created_at:
                          type: string
                          format: date-time
                        description:
                          type: string
                        id:
                          type: string
                        name:
                          type: string
                        project_id:
                          type: string
                        status:
                          type: string
                          enum:
                          - generating
                          - generated
                        user_id:
                          type: string
                  next_cursor:
                    oneOf:
                    - type: 'null'
                    - type: object
                      required: [created_at, podcast_id]
                      properties:
                        created_at:
                          type: string
                          format: date-time
                        podcast_id:
                          type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    post:
      tags: []
      operationId: create_podcast
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, description, used_file_ids]
              properties:
                description:
                  type: string
                  maxLength: 1024
                name:
                  type: string
                  maxLength: 256
                used_file_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [podcast_id]
                properties:
                  podcast_id:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
  /projects/{project_id}/podcasts/{podcast_id}:
    get:
      tags: []
      operationId: get_podcast
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      - name: podcast_id
        in: path
        description: Podcast ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [id, name, description, user_id, project_id, used_file_ids, podcast_script, audio_url, podcast_created_at]
                properties:
                  audio_url:
                    type: string
                  description:
                    type: string
                  id:
                    type: string
                  name:
                    type: string
                  podcast_created_at:
                    type: string
                    format: date-time
                  podcast_script:
                    type: array
                    items:
                      type: object
                      required: [speaker, text]
                      properties:
                        speaker:
                          type: string
                        text:
                          type: string
                  project_id:
                    type: string
                  used_file_ids:
                    type: array
                    items:
                      type: string
                  user_id:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    delete:
      tags: []
      operationId: remove_podcast
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      - name: podcast_id
        in: path
        description: Podcast ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Deleted
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    patch:
      tags: []
      operationId: update_podcast
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      - name: podcast_id
        in: path
        description: Podcast ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type:
                  - string
                  - 'null'
                  maxLength: 1024
                name:
                  type:
                  - string
                  - 'null'
                  maxLength: 256
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [id, name, description, user_id, project_id, used_file_ids, podcast_script, audio_url, podcast_created_at]
                properties:
                  audio_url:
                    type: string
                  description:
                    type: string
                  id:
                    type: string
                  name:
                    type: string
                  podcast_created_at:
                    type: string
                    format: date-time
                  podcast_script:
                    type: array
                    items:
                      type: object
                      required: [speaker, text]
                      properties:
                        speaker:
                          type: string
                        text:
                          type: string
                  project_id:
                    type: string
                  used_file_ids:
                    type: array
                    items:
                      type: string
                  user_id:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
...
components:
  schemas:
    ErrorBody:
      type: object
      required: [code, description]
      properties:
        code:
          type: string
        description:
          type: string